requirements.txt
requests
pyyaml
websockets
